# PHASE_1/API_Documentation/swagger.yaml
openapi: "3.0.0"
info:
  description: "This is the documentation for the interfaces and routes to the Viral API. https://github.com/aimenhamed/SENG3011_Viral"
  version: "1.0.0"
  title: "Viral API"
servers:
  - url: https://teamviral-api.herokuapp.com/api/v1/
    description: Production server (uses live data)
...
paths:
...
  /articles/dump:
    get:
      tags:
        - "Articles"
      summary: "Retrieve all articles"
      description: "Dumps all articles currently stored in the database. Main text is excluded."
      operationId: "articlesDump"
      responses:
        "200":
          description: "Successful response"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: "object"
                  properties:
                    articleId:
                      type: "string"
                    url:
                      type: "string"
                    dateOfPublication:
                      type: "string"
                    headline:
                      type: "string"
                    reports:
                      type: "array"
                      items:
                        "$ref": "#/components/schemas/Report"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "500":
          "$ref": "#/components/responses/ServerError"
  /articles/{articleId}:
    get:
      tags:
        - "Articles"
      summary: "Retrieves an article matching id"
      description: "Queries the database for an article with identical id"
      operationId: "articlesOne"
      parameters:
        - name: articleId
          in: path
          description: "a UUID representing the article being queried"
          schema:
            type: string
          required: true
          example: "c4ebd485-a22a-4ac3-882b-eec4daabd2b0"
      responses:
        "200":
          description: "Successful response"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  "$ref": "#/components/schemas/Article"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/ServerError"
...
components:
  responses:
    BadRequest:
      description: "The request was not formed with correct inputs."
      content:
        application/json:
          schema:
            type: "object"
            properties:
              errorCode:
                type: "integer"
                example: 400
              errorMessage:
                type: "string"
                example: "Bad request"
    NotFound:
      description: "The requested resource was not found"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              errorCode:
                type: "integer"
                example: 404
              errorMessage:
                type: "string"
                example: "Not found"
    ServerError:
      description: "Internal server error"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              errorCode:
                type: "integer"
                example: 500
              errorMessage:
                type: "string"
                example: "Internal server error"
  schemas:
...
    Report:
      type: "object"
      properties:
        reportId:
          type: "string"
        diseases:
          type: "array"
          items:
            type: "string"
        syndromes:
          type: "array"
          items:
            type: "string"
        eventDate:
          type: "string"
        locations:
          type: "array"
          items:
            type: "string"
    Article:
      type: "object"
      properties:
        articleId:
          type: "string"
        url:
          type: "string"
        dateOfPublication:
          type: "string"
        headline:
          type: "string"
        mainText:
          type: "string"
        reports:
          type: "array"
          items:
            "$ref": "#/components/schemas/Report"
...